+++ src/main.tsx
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom/client";
 import { Global } from "@emotion/react";
 
 import App from "@/App";
-import { GlobalStyle } from "@/change/GlobalStyle";
+import { GlobalStyle } from "@/Styles/GlobalStyle";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
 	<React.StrictMode>
